+++ ViewModels/WorkoutTypeViewModel.cs
@@ -4,7 +4,7 @@ namespace workout_app.ViewModels
 {
     public class WorkoutTypeViewModel
     {        
-        public long WorkoutTypeId { get; set; }
+        public int WorkoutTypeId { get; set; }
         public string WorkoutTypeName { get; set; }
     }
 }
